feat: move aws-sdk to dev deps
