+++ import-service/package.json
@@ -7,7 +7,6 @@
     "test": "jest --detectOpenHandles"
   },
   "dependencies": {
-    "aws-sdk": "^2.791.0",
     "csv-parser": "^2.3.3",
     "source-map-support": "^0.5.10"
   },
@@ -17,6 +16,7 @@
     "@types/jest": "^26.0.15",
     "@types/node": "^10.12.18",
     "@types/serverless": "^1.72.5",
+    "aws-sdk": "^2.792.0",
     "aws-sdk-mock": "^5.1.0",
     "fork-ts-checker-webpack-plugin": "^3.0.1",
     "jest": "^26.6.3",
